Extracts a history parameter as a matrix

Extracts a history parameter as a matrix with samplesN rows and randomN (found in object$data$randomN) columns.

Usage

extract_history_parameter(
  object,
  param_name,
  samplesN = NULL,
  link_function = NULL
)

Arguments

object

A cumhist object

param_name

String, a name of the parameter

samplesN

Number of samples, if NULL is computed from rstan (but it is cheaper to do this once).

link_function

A link function to use (exp or inv.logit) or NULL for identity.

Value

Matrix with samplesN rows and randomN (found in object$data$randomN) columns

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 5.666 seconds (Warm-up)
#> Chain 1:                5.98 seconds (Sampling)
#> Chain 1:                11.646 seconds (Total)
#> Chain 1: 
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess
extract_history_parameter(br_fit, "tau", link_function = exp)
#>              [,1]
#>    [1,] 1.0644964
#>    [2,] 0.9960436
#>    [3,] 0.8230219
#>    [4,] 0.8500291
#>    [5,] 1.0069592
#>    [6,] 0.8599762
#>    [7,] 1.2533189
#>    [8,] 0.9827817
#>    [9,] 1.1633156
#>   [10,] 1.0973242
#>   [11,] 0.9340265
#>   [12,] 1.0790112
#>   [13,] 0.7429560
#>   [14,] 0.8560725
#>   [15,] 1.0325606
#>   [16,] 1.2183829
#>   [17,] 0.9181795
#>   [18,] 1.1091461
#>   [19,] 1.0991949
#>   [20,] 1.0173111
#>   [21,] 0.9981620
#>   [22,] 1.0350662
#>   [23,] 0.9097209
#>   [24,] 0.8045297
#>   [25,] 0.9427634
#>   [26,] 0.8843964
#>   [27,] 1.0500814
#>   [28,] 1.0717855
#>   [29,] 1.1329443
#>   [30,] 1.0030932
#>   [31,] 0.8554808
#>   [32,] 0.7516949
#>   [33,] 1.1133038
#>   [34,] 0.9142508
#>   [35,] 1.0126275
#>   [36,] 1.1025390
#>   [37,] 1.0348263
#>   [38,] 0.8441943
#>   [39,] 0.9745436
#>   [40,] 0.9936197
#>   [41,] 0.8487997
#>   [42,] 0.9500149
#>   [43,] 0.9421060
#>   [44,] 1.1567039
#>   [45,] 0.9764469
#>   [46,] 1.1843307
#>   [47,] 0.9835433
#>   [48,] 0.7803060
#>   [49,] 1.1496373
#>   [50,] 0.9826454
#>   [51,] 0.8353241
#>   [52,] 1.0235374
#>   [53,] 1.0113475
#>   [54,] 1.0949637
#>   [55,] 1.2471327
#>   [56,] 1.0834217
#>   [57,] 0.9620807
#>   [58,] 1.0011907
#>   [59,] 0.9668201
#>   [60,] 1.1649485
#>   [61,] 0.8513478
#>   [62,] 0.9934275
#>   [63,] 1.0236907
#>   [64,] 1.0967518
#>   [65,] 0.9058658
#>   [66,] 1.1260260
#>   [67,] 1.1513270
#>   [68,] 1.0890223
#>   [69,] 0.9580381
#>   [70,] 0.9468607
#>   [71,] 1.1792605
#>   [72,] 0.8902806
#>   [73,] 0.8337704
#>   [74,] 1.1054506
#>   [75,] 1.1355083
#>   [76,] 0.9285187
#>   [77,] 1.1516658
#>   [78,] 1.1356716
#>   [79,] 0.8684089
#>   [80,] 1.1859882
#>   [81,] 0.9467046
#>   [82,] 0.9583658
#>   [83,] 1.0130980
#>   [84,] 1.1164523
#>   [85,] 0.8699467
#>   [86,] 0.9990963
#>   [87,] 1.1182319
#>   [88,] 0.7952341
#>   [89,] 0.9769800
#>   [90,] 0.9844688
#>   [91,] 0.8245958
#>   [92,] 1.0872056
#>   [93,] 0.7539712
#>   [94,] 0.8950974
#>   [95,] 0.7633295
#>   [96,] 0.9541209
#>   [97,] 1.0947522
#>   [98,] 1.2836274
#>   [99,] 0.8160129
#>  [100,] 0.9851578
#>  [101,] 0.9722595
#>  [102,] 1.0446334
#>  [103,] 0.9984763
#>  [104,] 0.9418748
#>  [105,] 0.8880803
#>  [106,] 0.8530196
#>  [107,] 1.1315249
#>  [108,] 0.9766202
#>  [109,] 1.0226429
#>  [110,] 0.9275043
#>  [111,] 0.9214337
#>  [112,] 0.9099607
#>  [113,] 1.0340367
#>  [114,] 1.0539994
#>  [115,] 0.7618016
#>  [116,] 0.9158671
#>  [117,] 0.7762242
#>  [118,] 0.9700616
#>  [119,] 0.9126060
#>  [120,] 0.9776330
#>  [121,] 1.0485191
#>  [122,] 1.0058762
#>  [123,] 0.9289261
#>  [124,] 0.9422056
#>  [125,] 0.9714289
#>  [126,] 1.1557023
#>  [127,] 0.8956034
#>  [128,] 1.0216114
#>  [129,] 0.8136244
#>  [130,] 0.9431378
#>  [131,] 0.9755966
#>  [132,] 1.2199630
#>  [133,] 0.8088533
#>  [134,] 0.8673136
#>  [135,] 0.8266101
#>  [136,] 0.8588465
#>  [137,] 1.1415861
#>  [138,] 0.7650248
#>  [139,] 0.8847474
#>  [140,] 0.8573863
#>  [141,] 1.1025521
#>  [142,] 0.7786324
#>  [143,] 0.9198140
#>  [144,] 1.1638929
#>  [145,] 0.9195984
#>  [146,] 0.8328266
#>  [147,] 1.0878080
#>  [148,] 0.9823897
#>  [149,] 1.0803510
#>  [150,] 0.8356064
#>  [151,] 0.9241223
#>  [152,] 0.9428769
#>  [153,] 1.0096576
#>  [154,] 0.9542590
#>  [155,] 0.8731789
#>  [156,] 0.8739028
#>  [157,] 0.9289878
#>  [158,] 1.1485479
#>  [159,] 0.8917264
#>  [160,] 0.8635636
#>  [161,] 1.1705939
#>  [162,] 1.0432903
#>  [163,] 1.1502798
#>  [164,] 0.8516790
#>  [165,] 1.0389536
#>  [166,] 1.0238172
#>  [167,] 0.9360604
#>  [168,] 1.0813434
#>  [169,] 0.8264520
#>  [170,] 0.9060209
#>  [171,] 0.9740551
#>  [172,] 1.2071280
#>  [173,] 0.8869579
#>  [174,] 0.8952473
#>  [175,] 1.0703719
#>  [176,] 0.9859153
#>  [177,] 0.8800780
#>  [178,] 0.8170354
#>  [179,] 0.7392583
#>  [180,] 0.9918889
#>  [181,] 1.3643428
#>  [182,] 1.1186670
#>  [183,] 1.0916920
#>  [184,] 0.7649281
#>  [185,] 0.9291941
#>  [186,] 0.7782021
#>  [187,] 0.8543643
#>  [188,] 0.9632669
#>  [189,] 1.0494223
#>  [190,] 0.8147305
#>  [191,] 1.0633419
#>  [192,] 1.0536475
#>  [193,] 1.0137288
#>  [194,] 0.8335696
#>  [195,] 0.8435674
#>  [196,] 1.0868848
#>  [197,] 0.9270063
#>  [198,] 1.0246200
#>  [199,] 0.8651446
#>  [200,] 0.9446560
#>  [201,] 0.9946831
#>  [202,] 1.2120100
#>  [203,] 0.8119596
#>  [204,] 1.4042295
#>  [205,] 0.8778305
#>  [206,] 1.0676841
#>  [207,] 0.9854594
#>  [208,] 0.8162944
#>  [209,] 1.0408275
#>  [210,] 1.0280329
#>  [211,] 0.8503689
#>  [212,] 0.9379891
#>  [213,] 1.3009080
#>  [214,] 0.9542243
#>  [215,] 0.9811914
#>  [216,] 1.1101650
#>  [217,] 0.9206168
#>  [218,] 0.7182010
#>  [219,] 0.8188293
#>  [220,] 0.9700987
#>  [221,] 0.8704067
#>  [222,] 0.8480782
#>  [223,] 0.8055231
#>  [224,] 0.8987296
#>  [225,] 1.2500228
#>  [226,] 0.8758267
#>  [227,] 0.9384572
#>  [228,] 0.9532942
#>  [229,] 0.9123850
#>  [230,] 1.1516294
#>  [231,] 0.8770495
#>  [232,] 1.1883289
#>  [233,] 1.1014744
#>  [234,] 1.0676788
#>  [235,] 1.2619573
#>  [236,] 0.7666006
#>  [237,] 0.7348492
#>  [238,] 0.9695952
#>  [239,] 0.7919957
#>  [240,] 1.0264675
#>  [241,] 0.9350116
#>  [242,] 0.9306518
#>  [243,] 1.1435175
#>  [244,] 1.2247848
#>  [245,] 1.1542359
#>  [246,] 0.9481316
#>  [247,] 0.7650470
#>  [248,] 0.7930653
#>  [249,] 0.8905161
#>  [250,] 1.0165685
#>  [251,] 0.9874286
#>  [252,] 0.9037953
#>  [253,] 1.0879866
#>  [254,] 0.9885428
#>  [255,] 0.9482439
#>  [256,] 1.0867165
#>  [257,] 0.9462783
#>  [258,] 0.9648485
#>  [259,] 1.0051250
#>  [260,] 1.2296803
#>  [261,] 0.9910454
#>  [262,] 0.8923384
#>  [263,] 1.1546554
#>  [264,] 1.2320298
#>  [265,] 1.0806848
#>  [266,] 0.9447362
#>  [267,] 1.0965378
#>  [268,] 1.1383253
#>  [269,] 0.9678125
#>  [270,] 1.1888224
#>  [271,] 0.9299785
#>  [272,] 0.8893411
#>  [273,] 0.9777944
#>  [274,] 1.1912424
#>  [275,] 0.7662472
#>  [276,] 0.8750370
#>  [277,] 0.8674487
#>  [278,] 0.8934471
#>  [279,] 1.1589874
#>  [280,] 0.8643381
#>  [281,] 1.0709254
#>  [282,] 1.0785410
#>  [283,] 1.1972913
#>  [284,] 1.0265247
#>  [285,] 0.7857863
#>  [286,] 0.9891501
#>  [287,] 1.1526235
#>  [288,] 1.2485208
#>  [289,] 0.8037958
#>  [290,] 1.0323631
#>  [291,] 0.7860092
#>  [292,] 1.0204177
#>  [293,] 0.8826771
#>  [294,] 0.7724837
#>  [295,] 0.7363722
#>  [296,] 0.9825250
#>  [297,] 1.0151776
#>  [298,] 1.0036801
#>  [299,] 0.9486908
#>  [300,] 0.9236656
#>  [301,] 1.0091828
#>  [302,] 0.9497575
#>  [303,] 1.1063321
#>  [304,] 0.9247874
#>  [305,] 1.0141559
#>  [306,] 1.0911475
#>  [307,] 1.1992003
#>  [308,] 0.7474278
#>  [309,] 0.9785369
#>  [310,] 0.8031919
#>  [311,] 0.9177063
#>  [312,] 1.2086591
#>  [313,] 1.3017202
#>  [314,] 1.0732590
#>  [315,] 0.7778946
#>  [316,] 1.1145625
#>  [317,] 1.2362490
#>  [318,] 1.0741602
#>  [319,] 1.2334812
#>  [320,] 0.7248766
#>  [321,] 1.0748259
#>  [322,] 0.8049957
#>  [323,] 1.1841354
#>  [324,] 1.0262682
#>  [325,] 0.9378966
#>  [326,] 0.9871178
#>  [327,] 1.0159766
#>  [328,] 1.0358164
#>  [329,] 0.9292950
#>  [330,] 1.0235185
#>  [331,] 0.7499460
#>  [332,] 0.8711979
#>  [333,] 0.8016313
#>  [334,] 1.0085429
#>  [335,] 0.9192662
#>  [336,] 1.0305212
#>  [337,] 1.1310341
#>  [338,] 0.9663516
#>  [339,] 0.9045115
#>  [340,] 0.8574126
#>  [341,] 1.1339388
#>  [342,] 0.8367180
#>  [343,] 0.7329975
#>  [344,] 0.9170199
#>  [345,] 0.9173196
#>  [346,] 0.7869458
#>  [347,] 0.9050986
#>  [348,] 0.7547299
#>  [349,] 1.0327615
#>  [350,] 0.9606654
#>  [351,] 0.8274406
#>  [352,] 1.0710950
#>  [353,] 0.9861769
#>  [354,] 1.0108115
#>  [355,] 0.9588047
#>  [356,] 1.0781401
#>  [357,] 1.0578124
#>  [358,] 0.9289031
#>  [359,] 0.8854609
#>  [360,] 0.7177209
#>  [361,] 0.9750122
#>  [362,] 1.0479767
#>  [363,] 1.0349706
#>  [364,] 0.9237908
#>  [365,] 0.9317550
#>  [366,] 0.9613342
#>  [367,] 0.9249542
#>  [368,] 0.9670758
#>  [369,] 0.9242933
#>  [370,] 0.9782864
#>  [371,] 1.0930859
#>  [372,] 0.8636122
#>  [373,] 1.0770298
#>  [374,] 1.1382066
#>  [375,] 0.9023302
#>  [376,] 1.1463226
#>  [377,] 1.0651701
#>  [378,] 1.0076542
#>  [379,] 1.1073257
#>  [380,] 1.1246642
#>  [381,] 0.9435688
#>  [382,] 0.7878384
#>  [383,] 0.8258944
#>  [384,] 0.9882795
#>  [385,] 0.9536485
#>  [386,] 1.1307461
#>  [387,] 1.0095845
#>  [388,] 0.9209415
#>  [389,] 1.1330704
#>  [390,] 1.1540527
#>  [391,] 1.0265581
#>  [392,] 0.8516159
#>  [393,] 0.9678774
#>  [394,] 0.9350538
#>  [395,] 0.9371868
#>  [396,] 1.0369329
#>  [397,] 0.9759114
#>  [398,] 0.8976366
#>  [399,] 1.2525408
#>  [400,] 0.7682241
#>  [401,] 0.9875067
#>  [402,] 1.0049732
#>  [403,] 0.7895771
#>  [404,] 1.0527124
#>  [405,] 1.1167821
#>  [406,] 0.7658424
#>  [407,] 1.1126774
#>  [408,] 0.7763765
#>  [409,] 1.0430265
#>  [410,] 1.0070526
#>  [411,] 1.1510101
#>  [412,] 0.9424039
#>  [413,] 0.8144382
#>  [414,] 1.0048647
#>  [415,] 0.8244123
#>  [416,] 1.1831651
#>  [417,] 0.7870858
#>  [418,] 0.8033039
#>  [419,] 1.1165342
#>  [420,] 1.0251405
#>  [421,] 1.1646322
#>  [422,] 0.9971288
#>  [423,] 0.8405719
#>  [424,] 0.9914756
#>  [425,] 0.8132438
#>  [426,] 1.1705882
#>  [427,] 1.0755073
#>  [428,] 1.3290954
#>  [429,] 0.9046366
#>  [430,] 0.7794996
#>  [431,] 0.8663054
#>  [432,] 1.1201367
#>  [433,] 1.2391268
#>  [434,] 0.8781362
#>  [435,] 0.8670868
#>  [436,] 1.0309563
#>  [437,] 0.9786251
#>  [438,] 1.0927044
#>  [439,] 0.9538393
#>  [440,] 0.9258300
#>  [441,] 0.9974645
#>  [442,] 0.8456301
#>  [443,] 1.0084775
#>  [444,] 0.9332716
#>  [445,] 1.0995818
#>  [446,] 0.7055367
#>  [447,] 0.7777109
#>  [448,] 0.7817665
#>  [449,] 1.0349222
#>  [450,] 1.0428352
#>  [451,] 1.0631685
#>  [452,] 0.8531252
#>  [453,] 1.1961137
#>  [454,] 0.8294552
#>  [455,] 0.9694766
#>  [456,] 1.0417644
#>  [457,] 0.8686346
#>  [458,] 0.7626943
#>  [459,] 1.1352822
#>  [460,] 0.8576598
#>  [461,] 0.9647736
#>  [462,] 1.1237151
#>  [463,] 0.7703935
#>  [464,] 1.0067198
#>  [465,] 1.0894116
#>  [466,] 1.0197013
#>  [467,] 1.1162130
#>  [468,] 0.9512640
#>  [469,] 1.1611685
#>  [470,] 0.7739124
#>  [471,] 0.9283914
#>  [472,] 0.9795492
#>  [473,] 1.2800450
#>  [474,] 0.7400073
#>  [475,] 1.0837466
#>  [476,] 1.0099246
#>  [477,] 0.9112756
#>  [478,] 1.2168647
#>  [479,] 1.0216680
#>  [480,] 1.0270679
#>  [481,] 0.9562060
#>  [482,] 1.0714244
#>  [483,] 1.4220592
#>  [484,] 0.9819019
#>  [485,] 0.9046139
#>  [486,] 0.9457210
#>  [487,] 0.9283934
#>  [488,] 1.0078369
#>  [489,] 1.2854844
#>  [490,] 0.8047569
#>  [491,] 0.8377491
#>  [492,] 0.8200450
#>  [493,] 0.8155119
#>  [494,] 1.0963944
#>  [495,] 1.3147280
#>  [496,] 0.9347550
#>  [497,] 1.0003714
#>  [498,] 1.1754415
#>  [499,] 1.3167478
#>  [500,] 1.2162964
#>  [501,] 1.2328234
#>  [502,] 1.0941001
#>  [503,] 0.9147443
#>  [504,] 0.8332372
#>  [505,] 0.8342597
#>  [506,] 0.8600063
#>  [507,] 1.0737886
#>  [508,] 0.9435953
#>  [509,] 0.9202688
#>  [510,] 1.1234409
#>  [511,] 1.1414850
#>  [512,] 0.8565144
#>  [513,] 0.9950971
#>  [514,] 1.2634001
#>  [515,] 0.9939462
#>  [516,] 0.8715971
#>  [517,] 0.9772980
#>  [518,] 1.1720930
#>  [519,] 1.1894715
#>  [520,] 1.0053353
#>  [521,] 0.8564155
#>  [522,] 1.0778612
#>  [523,] 1.0484153
#>  [524,] 1.0510051
#>  [525,] 1.1469912
#>  [526,] 1.1557150
#>  [527,] 1.0101390
#>  [528,] 0.8500522
#>  [529,] 1.1526077
#>  [530,] 0.8914982
#>  [531,] 1.1911156
#>  [532,] 0.8859553
#>  [533,] 0.6797687
#>  [534,] 1.3814133
#>  [535,] 0.7575560
#>  [536,] 1.0626446
#>  [537,] 0.9332493
#>  [538,] 0.8132537
#>  [539,] 0.8876322
#>  [540,] 1.1240035
#>  [541,] 1.1448090
#>  [542,] 0.9903687
#>  [543,] 0.7245736
#>  [544,] 0.9378330
#>  [545,] 0.9462916
#>  [546,] 0.9157446
#>  [547,] 1.1190801
#>  [548,] 0.9637636
#>  [549,] 0.8607720
#>  [550,] 1.1156533
#>  [551,] 0.9021372
#>  [552,] 1.1025628
#>  [553,] 1.1191214
#>  [554,] 0.7686797
#>  [555,] 1.0013189
#>  [556,] 1.0385301
#>  [557,] 1.0128885
#>  [558,] 0.8669667
#>  [559,] 0.8639220
#>  [560,] 0.9500617
#>  [561,] 0.9555696
#>  [562,] 1.1974159
#>  [563,] 0.9850605
#>  [564,] 1.0869866
#>  [565,] 0.7904638
#>  [566,] 0.9862868
#>  [567,] 0.9625965
#>  [568,] 0.9057442
#>  [569,] 0.8797552
#>  [570,] 0.8718212
#>  [571,] 1.0767672
#>  [572,] 0.9637598
#>  [573,] 0.9486572
#>  [574,] 1.0128412
#>  [575,] 0.9892642
#>  [576,] 0.9675013
#>  [577,] 1.0119144
#>  [578,] 0.8756446
#>  [579,] 0.9737250
#>  [580,] 1.3395609
#>  [581,] 1.0417943
#>  [582,] 0.9110374
#>  [583,] 1.2114058
#>  [584,] 1.1473021
#>  [585,] 1.1671210
#>  [586,] 0.8563270
#>  [587,] 0.9472644
#>  [588,] 0.9338995
#>  [589,] 1.0062240
#>  [590,] 0.8844981
#>  [591,] 1.2723778
#>  [592,] 1.0332096
#>  [593,] 0.9768724
#>  [594,] 0.8595884
#>  [595,] 0.8785922
#>  [596,] 0.9843690
#>  [597,] 0.9417086
#>  [598,] 0.9664546
#>  [599,] 0.9100384
#>  [600,] 0.8784775
#>  [601,] 0.8367823
#>  [602,] 0.8583464
#>  [603,] 0.8887963
#>  [604,] 1.1084178
#>  [605,] 0.9346394
#>  [606,] 1.0778110
#>  [607,] 1.2106262
#>  [608,] 1.0829542
#>  [609,] 0.9340218
#>  [610,] 0.7882597
#>  [611,] 0.9579363
#>  [612,] 0.9201514
#>  [613,] 0.9529735
#>  [614,] 1.0071793
#>  [615,] 1.1346783
#>  [616,] 0.8877876
#>  [617,] 0.8797161
#>  [618,] 0.9426895
#>  [619,] 0.9033455
#>  [620,] 0.8415231
#>  [621,] 0.9734729
#>  [622,] 0.9896746
#>  [623,] 0.9121836
#>  [624,] 1.1657436
#>  [625,] 1.0658448
#>  [626,] 0.9543476
#>  [627,] 0.9278676
#>  [628,] 0.8319251
#>  [629,] 1.0694705
#>  [630,] 1.1339951
#>  [631,] 0.8858814
#>  [632,] 0.9337372
#>  [633,] 1.0713848
#>  [634,] 0.9399761
#>  [635,] 1.1092291
#>  [636,] 0.9781418
#>  [637,] 0.8801138
#>  [638,] 0.7868170
#>  [639,] 1.1652170
#>  [640,] 1.1213615
#>  [641,] 0.9968469
#>  [642,] 1.0609677
#>  [643,] 1.0639649
#>  [644,] 0.9993383
#>  [645,] 0.7974377
#>  [646,] 0.9730071
#>  [647,] 0.8653758
#>  [648,] 0.7784178
#>  [649,] 0.9755292
#>  [650,] 1.0641685
#>  [651,] 1.1941307
#>  [652,] 1.2253388
#>  [653,] 0.9568750
#>  [654,] 1.0967669
#>  [655,] 0.9788519
#>  [656,] 0.7821466
#>  [657,] 1.0561330
#>  [658,] 1.2119085
#>  [659,] 0.7173837
#>  [660,] 1.0720524
#>  [661,] 0.8709622
#>  [662,] 0.7799197
#>  [663,] 0.9904431
#>  [664,] 1.0960772
#>  [665,] 0.7993931
#>  [666,] 0.7905612
#>  [667,] 1.0565267
#>  [668,] 1.1676958
#>  [669,] 0.9242393
#>  [670,] 0.9370541
#>  [671,] 0.8815185
#>  [672,] 0.9114511
#>  [673,] 0.8725962
#>  [674,] 0.8733600
#>  [675,] 1.0474633
#>  [676,] 1.1177692
#>  [677,] 0.7717983
#>  [678,] 1.0807877
#>  [679,] 0.8220763
#>  [680,] 0.9931054
#>  [681,] 0.8124361
#>  [682,] 0.7984894
#>  [683,] 1.1309783
#>  [684,] 1.0422433
#>  [685,] 1.1376934
#>  [686,] 0.8902348
#>  [687,] 1.0104682
#>  [688,] 0.9307173
#>  [689,] 1.1408495
#>  [690,] 0.9163101
#>  [691,] 0.9628954
#>  [692,] 1.0508868
#>  [693,] 1.0822675
#>  [694,] 0.9434722
#>  [695,] 1.0659331
#>  [696,] 0.9322564
#>  [697,] 1.0756962
#>  [698,] 1.4772438
#>  [699,] 0.9616350
#>  [700,] 0.8433620
#>  [701,] 1.0721619
#>  [702,] 0.9572289
#>  [703,] 1.1109488
#>  [704,] 1.0253017
#>  [705,] 1.0209441
#>  [706,] 1.1393045
#>  [707,] 0.9446106
#>  [708,] 0.8622058
#>  [709,] 1.1364490
#>  [710,] 1.0425507
#>  [711,] 1.0780414
#>  [712,] 1.1651444
#>  [713,] 1.0240463
#>  [714,] 0.9154325
#>  [715,] 0.8271892
#>  [716,] 0.8090642
#>  [717,] 0.9885725
#>  [718,] 1.0197084
#>  [719,] 1.0219151
#>  [720,] 1.0701485
#>  [721,] 1.0679014
#>  [722,] 0.8934435
#>  [723,] 0.9402385
#>  [724,] 1.1097054
#>  [725,] 1.1097400
#>  [726,] 0.8534073
#>  [727,] 1.0959995
#>  [728,] 0.7923450
#>  [729,] 0.9441592
#>  [730,] 0.8019843
#>  [731,] 0.9556923
#>  [732,] 0.7689038
#>  [733,] 0.8506250
#>  [734,] 0.8881846
#>  [735,] 1.1006493
#>  [736,] 1.1471113
#>  [737,] 1.1054019
#>  [738,] 1.1070216
#>  [739,] 0.8801733
#>  [740,] 1.0750278
#>  [741,] 1.0657105
#>  [742,] 0.8918011
#>  [743,] 0.7657938
#>  [744,] 0.8749249
#>  [745,] 1.1161069
#>  [746,] 0.9796143
#>  [747,] 0.9352828
#>  [748,] 1.0768345
#>  [749,] 1.1133624
#>  [750,] 0.9541606
#>  [751,] 0.9193258
#>  [752,] 0.8830779
#>  [753,] 1.0524498
#>  [754,] 1.4312877
#>  [755,] 1.1408295
#>  [756,] 0.7752101
#>  [757,] 1.0167069
#>  [758,] 1.0534016
#>  [759,] 0.7030489
#>  [760,] 0.8818757
#>  [761,] 0.9573647
#>  [762,] 0.8884515
#>  [763,] 0.9030808
#>  [764,] 0.8736986
#>  [765,] 0.9787062
#>  [766,] 0.8900988
#>  [767,] 1.0767873
#>  [768,] 0.8171701
#>  [769,] 0.9333081
#>  [770,] 1.0361401
#>  [771,] 0.9556175
#>  [772,] 0.9671898
#>  [773,] 0.8156145
#>  [774,] 0.9993928
#>  [775,] 0.9896495
#>  [776,] 1.0820051
#>  [777,] 0.7945853
#>  [778,] 0.9577153
#>  [779,] 0.9760221
#>  [780,] 1.2023250
#>  [781,] 0.9673740
#>  [782,] 0.8882379
#>  [783,] 1.2808789
#>  [784,] 1.0579261
#>  [785,] 0.9584766
#>  [786,] 0.9768863
#>  [787,] 0.9567960
#>  [788,] 1.3954797
#>  [789,] 0.9031585
#>  [790,] 1.1367593
#>  [791,] 0.7592126
#>  [792,] 1.1855529
#>  [793,] 1.1038407
#>  [794,] 0.9226300
#>  [795,] 0.9820851
#>  [796,] 0.8540621
#>  [797,] 1.1147299
#>  [798,] 1.2099607
#>  [799,] 1.2638991
#>  [800,] 0.9444427
#>  [801,] 0.8731020
#>  [802,] 0.8428490
#>  [803,] 0.9309619
#>  [804,] 0.8515938
#>  [805,] 1.1078930
#>  [806,] 0.9483237
#>  [807,] 1.3092476
#>  [808,] 0.9555188
#>  [809,] 1.0434646
#>  [810,] 0.8677967
#>  [811,] 0.8975052
#>  [812,] 1.2350066
#>  [813,] 1.1832141
#>  [814,] 1.0912697
#>  [815,] 0.8786651
#>  [816,] 1.1802556
#>  [817,] 0.9997761
#>  [818,] 1.0283237
#>  [819,] 1.0751476
#>  [820,] 1.1621728
#>  [821,] 0.8796320
#>  [822,] 1.1534975
#>  [823,] 1.1521143
#>  [824,] 1.0234950
#>  [825,] 0.8302094
#>  [826,] 1.1194667
#>  [827,] 0.8590939
#>  [828,] 1.0383962
#>  [829,] 0.9671301
#>  [830,] 1.0118804
#>  [831,] 1.0777867
#>  [832,] 0.9300983
#>  [833,] 0.9490566
#>  [834,] 0.7948762
#>  [835,] 0.8891419
#>  [836,] 0.9252264
#>  [837,] 0.7875442
#>  [838,] 1.0576789
#>  [839,] 1.0984919
#>  [840,] 1.0368720
#>  [841,] 1.1117642
#>  [842,] 0.9976047
#>  [843,] 1.1044067
#>  [844,] 1.1457548
#>  [845,] 1.0102531
#>  [846,] 0.9824804
#>  [847,] 0.8035122
#>  [848,] 1.2969094
#>  [849,] 1.2046737
#>  [850,] 0.7509912
#>  [851,] 0.8465392
#>  [852,] 1.0494275
#>  [853,] 1.0485819
#>  [854,] 0.8912908
#>  [855,] 0.9286151
#>  [856,] 0.9727813
#>  [857,] 0.9379492
#>  [858,] 1.1473590
#>  [859,] 0.9963303
#>  [860,] 0.7729862
#>  [861,] 1.0197192
#>  [862,] 0.9335764
#>  [863,] 0.8279961
#>  [864,] 0.9141928
#>  [865,] 0.8623137
#>  [866,] 0.8592586
#>  [867,] 0.8152158
#>  [868,] 1.0217109
#>  [869,] 1.0086641
#>  [870,] 1.2473172
#>  [871,] 0.7632532
#>  [872,] 0.8912582
#>  [873,] 0.9716137
#>  [874,] 0.8725044
#>  [875,] 0.9988815
#>  [876,] 0.8504251
#>  [877,] 0.9663205
#>  [878,] 0.8824298
#>  [879,] 1.0333707
#>  [880,] 0.9540832
#>  [881,] 0.8894562
#>  [882,] 0.8567155
#>  [883,] 1.0641599
#>  [884,] 0.8712263
#>  [885,] 0.9938430
#>  [886,] 0.9226532
#>  [887,] 1.1368500
#>  [888,] 1.0398418
#>  [889,] 1.0727859
#>  [890,] 0.7938929
#>  [891,] 0.8766771
#>  [892,] 1.0167152
#>  [893,] 0.8326852
#>  [894,] 1.2267530
#>  [895,] 0.9727010
#>  [896,] 0.8597765
#>  [897,] 1.1000598
#>  [898,] 1.2152017
#>  [899,] 0.8824817
#>  [900,] 0.8838340
#>  [901,] 1.0479865
#>  [902,] 0.8393248
#>  [903,] 1.0350844
#>  [904,] 1.0283613
#>  [905,] 0.7987114
#>  [906,] 0.9365128
#>  [907,] 0.9104385
#>  [908,] 0.9940481
#>  [909,] 0.8778923
#>  [910,] 0.8364829
#>  [911,] 0.8519052
#>  [912,] 1.0933607
#>  [913,] 0.8143084
#>  [914,] 1.0449954
#>  [915,] 1.1531685
#>  [916,] 0.9386542
#>  [917,] 0.8462519
#>  [918,] 1.0192954
#>  [919,] 0.9417132
#>  [920,] 1.0475192
#>  [921,] 1.0042855
#>  [922,] 1.2384760
#>  [923,] 1.2064411
#>  [924,] 0.6809973
#>  [925,] 1.0170555
#>  [926,] 0.9831414
#>  [927,] 1.0229351
#>  [928,] 0.9354371
#>  [929,] 0.9720062
#>  [930,] 1.0233275
#>  [931,] 0.8503376
#>  [932,] 1.0068106
#>  [933,] 0.7747047
#>  [934,] 1.0069018
#>  [935,] 0.9772676
#>  [936,] 0.9505398
#>  [937,] 1.2391216
#>  [938,] 1.0482216
#>  [939,] 1.1753096
#>  [940,] 1.1511425
#>  [941,] 0.8329854
#>  [942,] 0.8985835
#>  [943,] 0.9972231
#>  [944,] 1.0345269
#>  [945,] 0.7464593
#>  [946,] 0.9236145
#>  [947,] 0.8450773
#>  [948,] 0.7635340
#>  [949,] 0.8245044
#>  [950,] 0.9587403
#>  [951,] 0.9686015
#>  [952,] 1.1595343
#>  [953,] 0.8285553
#>  [954,] 0.9596193
#>  [955,] 0.8879619
#>  [956,] 1.0289656
#>  [957,] 0.8465588
#>  [958,] 1.1141169
#>  [959,] 1.0531825
#>  [960,] 0.9334763
#>  [961,] 0.8941001
#>  [962,] 0.9494101
#>  [963,] 0.9077426
#>  [964,] 0.7437924
#>  [965,] 1.1443753
#>  [966,] 1.0039831
#>  [967,] 0.9405682
#>  [968,] 0.9262141
#>  [969,] 1.1092539
#>  [970,] 0.9111281
#>  [971,] 1.0020443
#>  [972,] 0.9512418
#>  [973,] 0.9567531
#>  [974,] 0.8476751
#>  [975,] 0.9738218
#>  [976,] 0.8264467
#>  [977,] 1.0444977
#>  [978,] 0.9747068
#>  [979,] 1.1219858
#>  [980,] 1.0300285
#>  [981,] 0.8694738
#>  [982,] 1.0675826
#>  [983,] 0.9899005
#>  [984,] 0.9649526
#>  [985,] 0.9641293
#>  [986,] 0.9588028
#>  [987,] 1.1946726
#>  [988,] 0.8699311
#>  [989,] 1.0200620
#>  [990,] 1.1466958
#>  [991,] 1.0315619
#>  [992,] 0.8138703
#>  [993,] 1.1858885
#>  [994,] 1.1971052
#>  [995,] 0.9742069
#>  [996,] 0.9637449
#>  [997,] 0.9350503
#>  [998,] 1.0867579
#>  [999,] 0.9527127
#> [1000,] 0.8644189
# }