Extract values of used or fitted history parameter

Extract values of used or fitted history parameter

Usage

history_parameter(
  object,
  param,
  summary = TRUE,
  probs = c(0.055, 0.945),
  includePopulationLevel = TRUE
)

Arguments

object

An object of class cumhist

param

Parameter name: "tau" or "mixed_state"

summary

Whether summary statistics should be returned instead of raw sample values. Defaults to TRUE

probs

The percentiles used to compute summary, defaults to 89% credible interval.

includePopulationLevel

Logical, for pooled random effect only. Whether to include population mean as a separate "_population" level, default to TRUE.

Value

A vector, if summary was not requested. Or a tibble with a summary or if a fixed value was used.

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.848 seconds (Warm-up)
#> Chain 1:                4.601 seconds (Sampling)
#> Chain 1:                9.449 seconds (Total)
#> Chain 1: 
history_parameter(br_fit, "tau")
#> # A tibble: 1 x 3
#>   Estimate `5.5%` `94.5%`
#>      <dbl>  <dbl>   <dbl>
#> 1    0.990  0.787    1.23
# }