Computes history for a fitted model

Computes history for a fitted model, uses only mean values for each history parameter. Uses values for each random cluster, if "random" or "1|random" parametrisation was used.

Usage

extract_history(object)

Arguments

object

An object of class cumhist

Value

A matrix of cumulative history values for each state

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.474 seconds (Warm-up)
#> Chain 1:                4.458 seconds (Sampling)
#> Chain 1:                8.932 seconds (Total)
#> Chain 1: 
extract_history(br_fit)
#>               1         2
#>  [1,] 0.0000000 0.0000000
#>  [2,] 0.4401889 0.0000000
#>  [3,] 0.2249245 0.4890273
#>  [4,] 0.5389961 0.2908665
#>  [5,] 0.3173521 0.5824738
#>  [6,] 0.7961801 0.1739107
#>  [7,] 0.3117319 0.6765576
#>  [8,] 0.7613402 0.2345991
#>  [9,] 0.3418333 0.6563435
#> [10,] 0.6976691 0.3014934
#> [11,] 0.2544400 0.7452546
#> [12,] 0.8315354 0.1683956
#> [13,] 0.2585533 0.7414253
#> [14,] 0.7808584 0.2191353
#> [15,] 0.2906149 0.7093828
#> [16,] 0.7256809 0.2743182
#> [17,] 0.2619852 0.7380145
#> [18,] 0.6670392 0.3329606
#> [19,] 0.2847559 0.7152440
#> [20,] 0.6845131 0.3154869
#> [21,] 0.2673317 0.7326683
#> [22,] 0.7094014 0.2905986
#> [23,] 0.2367474 0.7632526
#> [24,] 0.7697906 0.2302094
#> [25,] 0.2836766 0.7163234
#> [26,] 0.7086619 0.2913381
#> [27,] 0.1803484 0.8196516
#> [28,] 0.7451418 0.2548582
#> [29,] 0.1735280 0.8264720
#> [30,] 0.8020378 0.1979622
#> [31,] 0.3164233 0.6835767
#> [32,] 0.7629670 0.2370330
#> [33,] 0.2897751 0.7102249
#> [34,] 0.7611073 0.2388927
#> [35,] 0.2017151 0.7982849
#> [36,] 0.6968112 0.3031888
#> [37,] 0.2391839 0.7608161
#> [38,] 0.8430859 0.1569141
#> [39,] 0.2326933 0.7673067
#> [40,] 0.7312891 0.2687109
#> [41,] 0.3861914 0.6138086
#> [42,] 0.6010337 0.3989663
#> [43,] 0.2377243 0.7622757
#> [44,] 0.8187982 0.1812018
#> [45,] 0.2926193 0.7073807
#> [46,] 0.8288361 0.1711639
#> [47,] 0.4375947 0.5624053
#> [48,] 0.5860840 0.4139160
#> [49,] 0.5576820 0.4423180
#> [50,] 0.3445605 0.6554395
#> [51,] 0.4302094 0.5697906
#> [52,] 0.6434666 0.3565334
#> [53,] 0.2538631 0.7461369
#> [54,] 0.6987596 0.3012404
#> [55,] 0.1764807 0.8235193
#> [56,] 0.8588545 0.1411455
#> [57,] 0.3133037 0.6866963
#> [58,] 0.4336666 0.5663334
#> [59,] 0.4701414 0.5298586
#> [60,] 0.6485424 0.3514576
#> [61,] 0.2289120 0.7710880
#> [62,] 0.6207957 0.3792043
#> [63,] 0.3022896 0.6977104
#> [64,] 0.7808460 0.2191540
#> [65,] 0.2106525 0.7893475
#> [66,] 0.6518268 0.3481732
#> [67,] 0.2985846 0.7014154
#> [68,] 0.7968014 0.2031986
#> [69,] 0.2090441 0.7909559
#> [70,] 0.8046916 0.1953084
#> [71,] 0.1429025 0.8570975
#> [72,] 0.8530994 0.1469006
#> [73,] 0.2755416 0.7244584
#> [74,] 0.7660080 0.2339920
#> [75,] 0.2135749 0.7864251
#> [76,] 0.8191653 0.1808347
# }