Computes history for a fitted model

Computes history for a fitted model, uses only mean values for each history parameter. Uses values for each random cluster, if "random" or "1|random" parametrisation was used.

Usage

extract_history(object)

Arguments

object

An object of class cumhist

Value

A matrix of cumulative history values for each state

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.862 seconds (Warm-up)
#> Chain 1:                6.093 seconds (Sampling)
#> Chain 1:                10.955 seconds (Total)
#> Chain 1: 
extract_history(br_fit)
#>               1         2
#>  [1,] 0.0000000 0.0000000
#>  [2,] 0.4417762 0.0000000
#>  [3,] 0.2249950 0.4907037
#>  [4,] 0.5402086 0.2911224
#>  [5,] 0.3172425 0.5837051
#>  [6,] 0.7973498 0.1732503
#>  [7,] 0.3107604 0.6777812
#>  [8,] 0.7622390 0.2338083
#>  [9,] 0.3408983 0.6573340
#> [10,] 0.6983902 0.3008009
#> [11,] 0.2534486 0.7462578
#> [12,] 0.8325349 0.1673992
#> [13,] 0.2573883 0.7425914
#> [14,] 0.7818195 0.2181745
#> [15,] 0.2895684 0.7104294
#> [16,] 0.7265508 0.2734484
#> [17,] 0.2609945 0.7390052
#> [18,] 0.6678886 0.3321113
#> [19,] 0.2839332 0.7160668
#> [20,] 0.6854129 0.3145871
#> [21,] 0.2664542 0.7335458
#> [22,] 0.7103672 0.2896328
#> [23,] 0.2357998 0.7642002
#> [24,] 0.7708530 0.2291470
#> [25,] 0.2826835 0.7173165
#> [26,] 0.7095397 0.2904603
#> [27,] 0.1793664 0.8206336
#> [28,] 0.7462912 0.2537088
#> [29,] 0.1725605 0.8274395
#> [30,] 0.8031875 0.1968125
#> [31,] 0.3154377 0.6845623
#> [32,] 0.7638525 0.2361475
#> [33,] 0.2887401 0.7112599
#> [34,] 0.7620315 0.2379685
#> [35,] 0.2006518 0.7993482
#> [36,] 0.6978424 0.3021576
#> [37,] 0.2382875 0.7617125
#> [38,] 0.8441102 0.1558898
#> [39,] 0.2315127 0.7684873
#> [40,] 0.7322541 0.2677459
#> [41,] 0.3854946 0.6145054
#> [42,] 0.6014221 0.3985779
#> [43,] 0.2368005 0.7631995
#> [44,] 0.8198498 0.1801502
#> [45,] 0.2915233 0.7084767
#> [46,] 0.8297573 0.1702427
#> [47,] 0.4367137 0.5632863
#> [48,] 0.5860571 0.4139429
#> [49,] 0.5575511 0.4424489
#> [50,] 0.3436688 0.6563312
#> [51,] 0.4300836 0.5699164
#> [52,] 0.6442052 0.3557948
#> [53,] 0.2530002 0.7469998
#> [54,] 0.6997471 0.3002529
#> [55,] 0.1755439 0.8244561
#> [56,] 0.8599085 0.1400915
#> [57,] 0.3121438 0.6878562
#> [58,] 0.4332448 0.5667552
#> [59,] 0.4700687 0.5299313
#> [60,] 0.6491997 0.3508003
#> [61,] 0.2279791 0.7720209
#> [62,] 0.6216534 0.3783466
#> [63,] 0.3016430 0.6983570
#> [64,] 0.7818827 0.2181173
#> [65,] 0.2095839 0.7904161
#> [66,] 0.6527493 0.3472507
#> [67,] 0.2978669 0.7021331
#> [68,] 0.7978232 0.2021768
#> [69,] 0.2079459 0.7920541
#> [70,] 0.8057547 0.1942453
#> [71,] 0.1418860 0.8581140
#> [72,] 0.8541893 0.1458107
#> [73,] 0.2743718 0.7256282
#> [74,] 0.7669229 0.2330771
#> [75,] 0.2124976 0.7875024
#> [76,] 0.8202156 0.1797844
# }