Computes history for a fitted model

Computes history for a fitted model, uses only mean values for each history parameter. Uses values for each random cluster, if "random" or "1|random" parametrisation was used.

Usage

extract_history(object)

Arguments

object

An object of class cumhist

Value

A matrix of cumulative history values for each state

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 6.996 seconds (Warm-up)
#> Chain 1:                6.767 seconds (Sampling)
#> Chain 1:                13.763 seconds (Total)
#> Chain 1: 
extract_history(br_fit)
#>               1         2
#>  [1,] 0.0000000 0.0000000
#>  [2,] 0.4406667 0.0000000
#>  [3,] 0.2249462 0.4895320
#>  [4,] 0.5393614 0.2909441
#>  [5,] 0.3173197 0.5828449
#>  [6,] 0.7965328 0.1737121
#>  [7,] 0.3114398 0.6769261
#>  [8,] 0.7616109 0.2343613
#>  [9,] 0.3415520 0.6566416
#> [10,] 0.6978861 0.3012851
#> [11,] 0.2541415 0.7455567
#> [12,] 0.8318365 0.1680954
#> [13,] 0.2582025 0.7417763
#> [14,] 0.7811477 0.2188461
#> [15,] 0.2902999 0.7096978
#> [16,] 0.7259427 0.2740565
#> [17,] 0.2616869 0.7383128
#> [18,] 0.6672947 0.3327052
#> [19,] 0.2845083 0.7154916
#> [20,] 0.6847838 0.3152161
#> [21,] 0.2670676 0.7329324
#> [22,] 0.7096921 0.2903079
#> [23,] 0.2364622 0.7635378
#> [24,] 0.7701106 0.2298894
#> [25,] 0.2833777 0.7166223
#> [26,] 0.7089260 0.2910740
#> [27,] 0.1800525 0.8199475
#> [28,] 0.7454878 0.2545122
#> [29,] 0.1732366 0.8267634
#> [30,] 0.8023842 0.1976158
#> [31,] 0.3161268 0.6838732
#> [32,] 0.7632336 0.2367664
#> [33,] 0.2894636 0.7105364
#> [34,] 0.7613855 0.2386145
#> [35,] 0.2013948 0.7986052
#> [36,] 0.6971214 0.3028786
#> [37,] 0.2389140 0.7610860
#> [38,] 0.8433946 0.1566054
#> [39,] 0.2323378 0.7676622
#> [40,] 0.7315795 0.2684205
#> [41,] 0.3859818 0.6140182
#> [42,] 0.6011505 0.3988495
#> [43,] 0.2374461 0.7625539
#> [44,] 0.8191150 0.1808850
#> [45,] 0.2922895 0.7077105
#> [46,] 0.8291136 0.1708864
#> [47,] 0.4373299 0.5626701
#> [48,] 0.5860759 0.4139241
#> [49,] 0.5576426 0.4423574
#> [50,] 0.3442920 0.6557080
#> [51,] 0.4301712 0.5698288
#> [52,] 0.6436887 0.3563113
#> [53,] 0.2536033 0.7463967
#> [54,] 0.6990568 0.3009432
#> [55,] 0.1761985 0.8238015
#> [56,] 0.8591722 0.1408278
#> [57,] 0.3129547 0.6870453
#> [58,] 0.4335395 0.5664605
#> [59,] 0.4701193 0.5298807
#> [60,] 0.6487401 0.3512599
#> [61,] 0.2286310 0.7713690
#> [62,] 0.6210536 0.3789464
#> [63,] 0.3020951 0.6979049
#> [64,] 0.7811583 0.2188417
#> [65,] 0.2103307 0.7896693
#> [66,] 0.6521042 0.3478958
#> [67,] 0.2983686 0.7016314
#> [68,] 0.7971092 0.2028908
#> [69,] 0.2087133 0.7912867
#> [70,] 0.8050118 0.1949882
#> [71,] 0.1425962 0.8574038
#> [72,] 0.8534279 0.1465721
#> [73,] 0.2751895 0.7248105
#> [74,] 0.7662834 0.2337166
#> [75,] 0.2132504 0.7867496
#> [76,] 0.8194817 0.1805183
# }