Computes history for a fitted model

Computes history for a fitted model, uses only mean values for each history parameter. Uses values for each random cluster, if "random" or "1|random" parametrisation was used.

Usage

extract_history(object)

Arguments

object

An object of class cumhist

Value

A matrix of cumulative history values for each state

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 5.33 seconds (Warm-up)
#> Chain 1:                5.106 seconds (Sampling)
#> Chain 1:                10.436 seconds (Total)
#> Chain 1: 
extract_history(br_fit)
#>               1         2
#>  [1,] 0.0000000 0.0000000
#>  [2,] 0.4406022 0.0000000
#>  [3,] 0.2249433 0.4894639
#>  [4,] 0.5393121 0.2909337
#>  [5,] 0.3173241 0.5827948
#>  [6,] 0.7964852 0.1737389
#>  [7,] 0.3114793 0.6768764
#>  [8,] 0.7615743 0.2343934
#>  [9,] 0.3415900 0.6566014
#> [10,] 0.6978568 0.3013132
#> [11,] 0.2541818 0.7455159
#> [12,] 0.8317959 0.1681359
#> [13,] 0.2582499 0.7417290
#> [14,] 0.7811087 0.2188851
#> [15,] 0.2903424 0.7096553
#> [16,] 0.7259073 0.2740918
#> [17,] 0.2617272 0.7382725
#> [18,] 0.6672602 0.3327396
#> [19,] 0.2845417 0.7154582
#> [20,] 0.6847473 0.3152527
#> [21,] 0.2671032 0.7328968
#> [22,] 0.7096528 0.2903472
#> [23,] 0.2365007 0.7634993
#> [24,] 0.7700674 0.2299326
#> [25,] 0.2834181 0.7165819
#> [26,] 0.7088904 0.2911096
#> [27,] 0.1800925 0.8199075
#> [28,] 0.7454411 0.2545589
#> [29,] 0.1732759 0.8267241
#> [30,] 0.8023374 0.1976626
#> [31,] 0.3161669 0.6838331
#> [32,] 0.7631976 0.2368024
#> [33,] 0.2895056 0.7104944
#> [34,] 0.7613479 0.2386521
#> [35,] 0.2014381 0.7985619
#> [36,] 0.6970795 0.3029205
#> [37,] 0.2389504 0.7610496
#> [38,] 0.8433529 0.1566471
#> [39,] 0.2323858 0.7676142
#> [40,] 0.7315402 0.2684598
#> [41,] 0.3860101 0.6139899
#> [42,] 0.6011348 0.3988652
#> [43,] 0.2374837 0.7625163
#> [44,] 0.8190723 0.1809277
#> [45,] 0.2923340 0.7076660
#> [46,] 0.8290762 0.1709238
#> [47,] 0.4373656 0.5626344
#> [48,] 0.5860770 0.4139230
#> [49,] 0.5576479 0.4423521
#> [50,] 0.3443283 0.6556717
#> [51,] 0.4301763 0.5698237
#> [52,] 0.6436587 0.3563413
#> [53,] 0.2536383 0.7463617
#> [54,] 0.6990167 0.3009833
#> [55,] 0.1762366 0.8237634
#> [56,] 0.8591293 0.1408707
#> [57,] 0.3130018 0.6869982
#> [58,] 0.4335566 0.5664434
#> [59,] 0.4701223 0.5298777
#> [60,] 0.6487134 0.3512866
#> [61,] 0.2286689 0.7713311
#> [62,] 0.6210188 0.3789812
#> [63,] 0.3021213 0.6978787
#> [64,] 0.7811161 0.2188839
#> [65,] 0.2103741 0.7896259
#> [66,] 0.6520668 0.3479332
#> [67,] 0.2983978 0.7016022
#> [68,] 0.7970677 0.2029323
#> [69,] 0.2087580 0.7912420
#> [70,] 0.8049686 0.1950314
#> [71,] 0.1426375 0.8573625
#> [72,] 0.8533836 0.1466164
#> [73,] 0.2752371 0.7247629
#> [74,] 0.7662462 0.2337538
#> [75,] 0.2132942 0.7867058
#> [76,] 0.8194390 0.1805610
# }