Computes history for a fitted model

Computes history for a fitted model, uses only mean values for each history parameter. Uses values for each random cluster, if "random" or "1|random" parametrisation was used.

Usage

extract_history(object)

Arguments

object

An object of class cumhist

Value

A matrix of cumulative history values for each state

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 5.336 seconds (Warm-up)
#> Chain 1:                6.376 seconds (Sampling)
#> Chain 1:                11.712 seconds (Total)
#> Chain 1: 
extract_history(br_fit)
#>               1         2
#>  [1,] 0.0000000 0.0000000
#>  [2,] 0.4390140 0.0000000
#>  [3,] 0.2248691 0.4877860
#>  [4,] 0.5380967 0.2906734
#>  [5,] 0.3174296 0.5815599
#>  [6,] 0.7953111 0.1743979
#>  [7,] 0.3124486 0.6756512
#>  [8,] 0.7606745 0.2351832
#>  [9,] 0.3425241 0.6556106
#> [10,] 0.6971363 0.3020045
#> [11,] 0.2551741 0.7445114
#> [12,] 0.8307942 0.1691343
#> [13,] 0.2594156 0.7405621
#> [14,] 0.7801470 0.2198464
#> [15,] 0.2913890 0.7086085
#> [16,] 0.7250378 0.2749612
#> [17,] 0.2627185 0.7372812
#> [18,] 0.6664120 0.3335879
#> [19,] 0.2853646 0.7146353
#> [20,] 0.6838478 0.3161522
#> [21,] 0.2679808 0.7320192
#> [22,] 0.7086870 0.2913130
#> [23,] 0.2374488 0.7625512
#> [24,] 0.7690038 0.2309962
#> [25,] 0.2844108 0.7155892
#> [26,] 0.7080128 0.2919872
#> [27,] 0.1810764 0.8189236
#> [28,] 0.7442910 0.2557090
#> [29,] 0.1742450 0.8257550
#> [30,] 0.8011855 0.1988145
#> [31,] 0.3171512 0.6828488
#> [32,] 0.7623114 0.2376886
#> [33,] 0.2905408 0.7094592
#> [34,] 0.7604234 0.2395766
#> [35,] 0.2025033 0.7974967
#> [36,] 0.6960490 0.3039510
#> [37,] 0.2398472 0.7601528
#> [38,] 0.8423258 0.1576742
#> [39,] 0.2335676 0.7664324
#> [40,] 0.7305756 0.2694244
#> [41,] 0.3867058 0.6132942
#> [42,] 0.6007470 0.3992530
#> [43,] 0.2384087 0.7615913
#> [44,] 0.8180186 0.1819814
#> [45,] 0.2934297 0.7065703
#> [46,] 0.8281533 0.1718467
#> [47,] 0.4382446 0.5617554
#> [48,] 0.5861040 0.4138960
#> [49,] 0.5577789 0.4422211
#> [50,] 0.3452206 0.6547794
#> [51,] 0.4303044 0.5696956
#> [52,] 0.6429215 0.3570785
#> [53,] 0.2545020 0.7454980
#> [54,] 0.6980295 0.3019705
#> [55,] 0.1771752 0.8228248
#> [56,] 0.8580717 0.1419283
#> [57,] 0.3141609 0.6858391
#> [58,] 0.4339795 0.5660205
#> [59,] 0.4701964 0.5298036
#> [60,] 0.6480572 0.3519428
#> [61,] 0.2296031 0.7703969
#> [62,] 0.6201626 0.3798374
#> [63,] 0.3027677 0.6972323
#> [64,] 0.7800778 0.2199222
#> [65,] 0.2114439 0.7885561
#> [66,] 0.6511456 0.3488544
#> [67,] 0.2991153 0.7008847
#> [68,] 0.7960441 0.2039559
#> [69,] 0.2098577 0.7901423
#> [70,] 0.8039039 0.1960961
#> [71,] 0.1436570 0.8563430
#> [72,] 0.8522902 0.1477098
#> [73,] 0.2764068 0.7235932
#> [74,] 0.7653310 0.2346690
#> [75,] 0.2143733 0.7856267
#> [76,] 0.8183867 0.1816133
# }