Extracts a history parameter as a matrix

Extracts a history parameter as a matrix with samplesN rows and randomN (found in object$data$randomN) columns.

Usage

extract_history_parameter(
  object,
  param_name,
  samplesN = NULL,
  link_function = NULL
)

Arguments

object

A cumhist object

param_name

String, a name of the parameter

samplesN

Number of samples, if NULL is computed from rstan (but it is cheaper to do this once).

link_function

A link function to use (exp or inv.logit) or NULL for identity.

Value

Matrix with samplesN rows and randomN (found in object$data$randomN) columns

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state="State", duration="Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 4.538 seconds (Warm-up)
#> Chain 1:                4.461 seconds (Sampling)
#> Chain 1:                8.999 seconds (Total)
#> Chain 1: 
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
extract_history_parameter(br_fit, "tau", link_function = exp)
#>              [,1]
#>    [1,] 1.1977910
#>    [2,] 0.8903915
#>    [3,] 0.8994296
#>    [4,] 1.1215123
#>    [5,] 1.2533425
#>    [6,] 0.8218354
#>    [7,] 0.9612333
#>    [8,] 0.9582499
#>    [9,] 1.0244239
#>   [10,] 1.0501708
#>   [11,] 0.9210404
#>   [12,] 1.2974571
#>   [13,] 0.8331778
#>   [14,] 1.0137403
#>   [15,] 0.7809877
#>   [16,] 0.7363882
#>   [17,] 0.9181438
#>   [18,] 1.2143905
#>   [19,] 1.0479115
#>   [20,] 0.9024283
#>   [21,] 0.9134785
#>   [22,] 0.9501740
#>   [23,] 1.0210785
#>   [24,] 1.0206202
#>   [25,] 1.0800878
#>   [26,] 1.0422466
#>   [27,] 1.3025811
#>   [28,] 1.0697167
#>   [29,] 0.8762928
#>   [30,] 1.1253990
#>   [31,] 1.0456162
#>   [32,] 1.0035757
#>   [33,] 0.9484325
#>   [34,] 1.0678294
#>   [35,] 1.0386519
#>   [36,] 1.0018750
#>   [37,] 1.1247503
#>   [38,] 1.0534238
#>   [39,] 0.8746860
#>   [40,] 1.0919495
#>   [41,] 0.7809390
#>   [42,] 0.9620999
#>   [43,] 1.0809869
#>   [44,] 0.7067100
#>   [45,] 0.9849882
#>   [46,] 1.3793939
#>   [47,] 0.8239488
#>   [48,] 1.0023433
#>   [49,] 0.9760186
#>   [50,] 1.1786625
#>   [51,] 1.0934917
#>   [52,] 0.9331085
#>   [53,] 1.0396375
#>   [54,] 0.9837384
#>   [55,] 0.9083071
#>   [56,] 1.1712175
#>   [57,] 1.1822377
#>   [58,] 0.9862242
#>   [59,] 0.9614715
#>   [60,] 0.9808263
#>   [61,] 1.2377113
#>   [62,] 0.9802327
#>   [63,] 0.8305467
#>   [64,] 0.9700041
#>   [65,] 0.8301804
#>   [66,] 1.0134055
#>   [67,] 0.9625571
#>   [68,] 1.2752886
#>   [69,] 0.8545650
#>   [70,] 1.2238082
#>   [71,] 1.0480948
#>   [72,] 0.9412943
#>   [73,] 1.1851548
#>   [74,] 0.7950686
#>   [75,] 1.1118284
#>   [76,] 0.8324737
#>   [77,] 0.9618627
#>   [78,] 0.9736496
#>   [79,] 1.1014166
#>   [80,] 1.0416275
#>   [81,] 0.8358296
#>   [82,] 0.9204869
#>   [83,] 1.0765272
#>   [84,] 0.8222010
#>   [85,] 1.0851916
#>   [86,] 1.0365725
#>   [87,] 1.1099516
#>   [88,] 0.9500440
#>   [89,] 0.8798126
#>   [90,] 0.9417696
#>   [91,] 0.8324340
#>   [92,] 0.8908673
#>   [93,] 1.0288280
#>   [94,] 0.8485888
#>   [95,] 1.4132468
#>   [96,] 1.0304835
#>   [97,] 0.8178029
#>   [98,] 0.7202939
#>   [99,] 1.3424446
#>  [100,] 1.0101968
#>  [101,] 0.9354804
#>  [102,] 1.0882392
#>  [103,] 1.0156331
#>  [104,] 0.8726135
#>  [105,] 1.1904173
#>  [106,] 1.0997768
#>  [107,] 1.0487179
#>  [108,] 0.9188702
#>  [109,] 0.8681958
#>  [110,] 0.9669903
#>  [111,] 1.0100150
#>  [112,] 0.9196950
#>  [113,] 1.0567713
#>  [114,] 1.3435454
#>  [115,] 1.0843529
#>  [116,] 1.2112614
#>  [117,] 0.8906707
#>  [118,] 0.9318579
#>  [119,] 0.8544818
#>  [120,] 1.2865188
#>  [121,] 1.0759331
#>  [122,] 0.8005639
#>  [123,] 0.8229885
#>  [124,] 0.7801536
#>  [125,] 0.9517744
#>  [126,] 1.0619455
#>  [127,] 1.0998796
#>  [128,] 1.0053138
#>  [129,] 0.7939793
#>  [130,] 0.7886777
#>  [131,] 1.0004184
#>  [132,] 0.9451247
#>  [133,] 0.9702420
#>  [134,] 1.0049385
#>  [135,] 1.0407745
#>  [136,] 1.0550943
#>  [137,] 1.0055454
#>  [138,] 0.8725860
#>  [139,] 1.0900084
#>  [140,] 1.0593060
#>  [141,] 0.9391955
#>  [142,] 0.8930322
#>  [143,] 0.9380787
#>  [144,] 0.9390266
#>  [145,] 1.0125935
#>  [146,] 0.9121218
#>  [147,] 0.6622459
#>  [148,] 1.1301232
#>  [149,] 0.8374901
#>  [150,] 1.0100593
#>  [151,] 0.8930363
#>  [152,] 1.0734250
#>  [153,] 0.8986424
#>  [154,] 1.0574446
#>  [155,] 0.8907056
#>  [156,] 1.0247263
#>  [157,] 0.8759759
#>  [158,] 0.9017612
#>  [159,] 0.8572273
#>  [160,] 0.9955234
#>  [161,] 1.0504789
#>  [162,] 0.8616091
#>  [163,] 1.0309425
#>  [164,] 0.9808341
#>  [165,] 1.1957920
#>  [166,] 1.0837970
#>  [167,] 0.7807291
#>  [168,] 0.7969647
#>  [169,] 0.7789168
#>  [170,] 1.0033406
#>  [171,] 1.0194235
#>  [172,] 1.0204595
#>  [173,] 0.9240134
#>  [174,] 1.0570770
#>  [175,] 0.7602548
#>  [176,] 0.7001906
#>  [177,] 0.9242979
#>  [178,] 0.9768601
#>  [179,] 0.8345593
#>  [180,] 1.1624867
#>  [181,] 0.9428350
#>  [182,] 0.7615137
#>  [183,] 1.0053923
#>  [184,] 0.8339227
#>  [185,] 0.9295023
#>  [186,] 0.9170846
#>  [187,] 1.0152001
#>  [188,] 0.8564736
#>  [189,] 0.9271068
#>  [190,] 0.8211275
#>  [191,] 0.9507858
#>  [192,] 1.1951290
#>  [193,] 0.8194661
#>  [194,] 0.9911759
#>  [195,] 0.9205488
#>  [196,] 1.0136891
#>  [197,] 0.8634828
#>  [198,] 1.2725174
#>  [199,] 1.1751237
#>  [200,] 1.2128725
#>  [201,] 0.9282886
#>  [202,] 1.1902700
#>  [203,] 0.8575585
#>  [204,] 1.1217641
#>  [205,] 0.9171877
#>  [206,] 0.8214717
#>  [207,] 0.8247037
#>  [208,] 0.8708035
#>  [209,] 0.8107018
#>  [210,] 1.0313824
#>  [211,] 0.9366605
#>  [212,] 0.8247603
#>  [213,] 0.8910275
#>  [214,] 1.1344844
#>  [215,] 0.8462223
#>  [216,] 1.0050329
#>  [217,] 1.0093156
#>  [218,] 0.8946599
#>  [219,] 0.9659117
#>  [220,] 0.9684481
#>  [221,] 0.9715180
#>  [222,] 0.9861623
#>  [223,] 0.8194802
#>  [224,] 0.9721279
#>  [225,] 0.8595587
#>  [226,] 1.1767140
#>  [227,] 1.0598704
#>  [228,] 0.9045282
#>  [229,] 0.9720274
#>  [230,] 1.2510891
#>  [231,] 1.0094815
#>  [232,] 0.9343438
#>  [233,] 1.1521773
#>  [234,] 0.9909731
#>  [235,] 0.7146101
#>  [236,] 1.0453321
#>  [237,] 1.4785739
#>  [238,] 0.9462207
#>  [239,] 1.2150121
#>  [240,] 0.9762918
#>  [241,] 1.0839210
#>  [242,] 0.8990466
#>  [243,] 0.9238851
#>  [244,] 1.2116524
#>  [245,] 1.0842964
#>  [246,] 0.8429592
#>  [247,] 0.8724511
#>  [248,] 0.8944542
#>  [249,] 1.0850532
#>  [250,] 1.1180536
#>  [251,] 1.1373389
#>  [252,] 0.8377051
#>  [253,] 0.9541024
#>  [254,] 1.0030810
#>  [255,] 0.8904361
#>  [256,] 1.0746880
#>  [257,] 1.1223683
#>  [258,] 0.8333735
#>  [259,] 1.0376052
#>  [260,] 1.0023470
#>  [261,] 0.8029743
#>  [262,] 0.9625305
#>  [263,] 0.9407723
#>  [264,] 0.9384498
#>  [265,] 1.1203330
#>  [266,] 0.8720858
#>  [267,] 1.0583967
#>  [268,] 0.9476381
#>  [269,] 1.0620116
#>  [270,] 0.9576081
#>  [271,] 0.8685551
#>  [272,] 1.1426480
#>  [273,] 0.8639542
#>  [274,] 0.9562522
#>  [275,] 1.0724908
#>  [276,] 0.7662798
#>  [277,] 0.7357866
#>  [278,] 0.9949245
#>  [279,] 1.1728448
#>  [280,] 1.0043746
#>  [281,] 0.8924925
#>  [282,] 1.3847595
#>  [283,] 0.9492167
#>  [284,] 1.3002559
#>  [285,] 1.0596409
#>  [286,] 1.3034121
#>  [287,] 0.8259463
#>  [288,] 1.0411851
#>  [289,] 1.1606759
#>  [290,] 0.8348156
#>  [291,] 1.0364709
#>  [292,] 1.1200970
#>  [293,] 0.7945548
#>  [294,] 0.8389617
#>  [295,] 1.1643523
#>  [296,] 0.9851354
#>  [297,] 0.9409040
#>  [298,] 0.8143515
#>  [299,] 0.9089769
#>  [300,] 1.0769809
#>  [301,] 1.1923965
#>  [302,] 0.7410829
#>  [303,] 1.1970900
#>  [304,] 1.0977107
#>  [305,] 0.7767250
#>  [306,] 0.9646532
#>  [307,] 0.9324466
#>  [308,] 0.9813010
#>  [309,] 1.0955175
#>  [310,] 0.9291725
#>  [311,] 1.0078559
#>  [312,] 1.1010279
#>  [313,] 0.9532632
#>  [314,] 0.8225930
#>  [315,] 0.8370225
#>  [316,] 0.9049467
#>  [317,] 0.9164605
#>  [318,] 1.0165076
#>  [319,] 0.8845835
#>  [320,] 1.3845802
#>  [321,] 0.7223954
#>  [322,] 0.9819295
#>  [323,] 1.0586948
#>  [324,] 1.0380059
#>  [325,] 0.9936395
#>  [326,] 0.8867197
#>  [327,] 0.7848292
#>  [328,] 1.0390963
#>  [329,] 0.9547143
#>  [330,] 1.2034687
#>  [331,] 1.1062365
#>  [332,] 1.1526776
#>  [333,] 0.9638711
#>  [334,] 1.0524364
#>  [335,] 1.3544018
#>  [336,] 1.0090959
#>  [337,] 0.8826302
#>  [338,] 0.7578957
#>  [339,] 0.7278327
#>  [340,] 0.8653527
#>  [341,] 0.9402800
#>  [342,] 1.0207617
#>  [343,] 1.2224972
#>  [344,] 0.9317835
#>  [345,] 1.1134942
#>  [346,] 0.9445722
#>  [347,] 0.8844957
#>  [348,] 0.9451993
#>  [349,] 1.0167867
#>  [350,] 1.0642729
#>  [351,] 0.7900098
#>  [352,] 1.3062006
#>  [353,] 1.1277751
#>  [354,] 1.1082328
#>  [355,] 0.9962461
#>  [356,] 0.9161805
#>  [357,] 0.8432617
#>  [358,] 0.8131377
#>  [359,] 1.0205197
#>  [360,] 0.9179734
#>  [361,] 0.8686950
#>  [362,] 1.1525456
#>  [363,] 1.0252772
#>  [364,] 0.9940488
#>  [365,] 1.0469731
#>  [366,] 0.8404387
#>  [367,] 0.9863486
#>  [368,] 1.4858358
#>  [369,] 1.1715457
#>  [370,] 0.9824671
#>  [371,] 0.9933189
#>  [372,] 0.9703036
#>  [373,] 0.8011342
#>  [374,] 0.9112214
#>  [375,] 1.3580846
#>  [376,] 1.0621712
#>  [377,] 1.1866928
#>  [378,] 1.0586596
#>  [379,] 1.1271019
#>  [380,] 1.1072042
#>  [381,] 1.0563483
#>  [382,] 1.1206362
#>  [383,] 1.0458183
#>  [384,] 0.8911549
#>  [385,] 0.8315556
#>  [386,] 0.9583693
#>  [387,] 0.8038233
#>  [388,] 1.1241913
#>  [389,] 0.9276520
#>  [390,] 0.8595967
#>  [391,] 0.9942766
#>  [392,] 1.0985997
#>  [393,] 1.3267093
#>  [394,] 1.2275688
#>  [395,] 0.9378172
#>  [396,] 0.9579262
#>  [397,] 0.9723583
#>  [398,] 0.8014846
#>  [399,] 0.8981737
#>  [400,] 0.8159792
#>  [401,] 1.0062370
#>  [402,] 0.9071274
#>  [403,] 0.8592457
#>  [404,] 0.8868043
#>  [405,] 1.3269175
#>  [406,] 1.1759509
#>  [407,] 1.0132581
#>  [408,] 0.9752542
#>  [409,] 0.7698420
#>  [410,] 1.1605533
#>  [411,] 1.0356055
#>  [412,] 1.1089667
#>  [413,] 0.8791373
#>  [414,] 1.1015818
#>  [415,] 0.8351268
#>  [416,] 1.0094934
#>  [417,] 0.8191160
#>  [418,] 0.8827900
#>  [419,] 0.9416114
#>  [420,] 0.7564200
#>  [421,] 1.0776886
#>  [422,] 1.0611153
#>  [423,] 0.9271997
#>  [424,] 0.8808414
#>  [425,] 1.0386578
#>  [426,] 0.8534122
#>  [427,] 1.0072007
#>  [428,] 0.8773489
#>  [429,] 0.8611318
#>  [430,] 1.0594262
#>  [431,] 1.2525904
#>  [432,] 0.8275928
#>  [433,] 0.8688343
#>  [434,] 1.0628594
#>  [435,] 1.0533123
#>  [436,] 1.1156404
#>  [437,] 1.1746845
#>  [438,] 0.8344028
#>  [439,] 1.2002023
#>  [440,] 0.8234816
#>  [441,] 1.2987077
#>  [442,] 0.9835087
#>  [443,] 0.8521741
#>  [444,] 0.8795096
#>  [445,] 1.0214936
#>  [446,] 1.1544959
#>  [447,] 0.9738338
#>  [448,] 1.1166953
#>  [449,] 0.9242893
#>  [450,] 0.8574258
#>  [451,] 1.0733605
#>  [452,] 0.7863622
#>  [453,] 0.8554940
#>  [454,] 0.9109881
#>  [455,] 0.9765255
#>  [456,] 0.8705426
#>  [457,] 1.0039013
#>  [458,] 0.9045206
#>  [459,] 1.0309040
#>  [460,] 1.2193808
#>  [461,] 0.9028147
#>  [462,] 1.2074753
#>  [463,] 0.8570050
#>  [464,] 0.9198163
#>  [465,] 0.7505352
#>  [466,] 0.8392495
#>  [467,] 0.9383464
#>  [468,] 0.8882284
#>  [469,] 0.9515326
#>  [470,] 1.2201789
#>  [471,] 0.9843742
#>  [472,] 0.8488440
#>  [473,] 1.2919531
#>  [474,] 1.1164054
#>  [475,] 1.0112835
#>  [476,] 1.0890844
#>  [477,] 0.8062807
#>  [478,] 1.1675572
#>  [479,] 0.9273665
#>  [480,] 1.1393228
#>  [481,] 0.8943885
#>  [482,] 0.9879690
#>  [483,] 0.9892215
#>  [484,] 0.9616299
#>  [485,] 1.0113668
#>  [486,] 0.9224569
#>  [487,] 1.2401539
#>  [488,] 1.0646019
#>  [489,] 1.0665684
#>  [490,] 1.1372104
#>  [491,] 0.8686779
#>  [492,] 1.0145689
#>  [493,] 0.9526842
#>  [494,] 1.2748987
#>  [495,] 0.8492377
#>  [496,] 0.9673115
#>  [497,] 1.0238315
#>  [498,] 0.9959070
#>  [499,] 1.3382613
#>  [500,] 1.1027696
#>  [501,] 0.9991192
#>  [502,] 1.2219279
#>  [503,] 1.1359146
#>  [504,] 1.0960148
#>  [505,] 1.0359974
#>  [506,] 1.2365169
#>  [507,] 0.9674481
#>  [508,] 1.2003046
#>  [509,] 0.8397812
#>  [510,] 0.8154142
#>  [511,] 0.9021898
#>  [512,] 0.9758127
#>  [513,] 0.7300946
#>  [514,] 0.9842370
#>  [515,] 0.9282697
#>  [516,] 1.3351055
#>  [517,] 0.8691533
#>  [518,] 1.0140466
#>  [519,] 1.1371417
#>  [520,] 1.2026510
#>  [521,] 0.7607563
#>  [522,] 0.8920017
#>  [523,] 0.8654338
#>  [524,] 1.1795223
#>  [525,] 1.0432035
#>  [526,] 1.0410953
#>  [527,] 0.9127306
#>  [528,] 1.0179670
#>  [529,] 0.8663067
#>  [530,] 0.9308198
#>  [531,] 1.0185937
#>  [532,] 0.9983343
#>  [533,] 0.9767884
#>  [534,] 0.9927309
#>  [535,] 1.1219991
#>  [536,] 1.1114814
#>  [537,] 0.9481037
#>  [538,] 0.9179206
#>  [539,] 1.1179345
#>  [540,] 0.9341861
#>  [541,] 1.3090262
#>  [542,] 1.1714332
#>  [543,] 0.9988258
#>  [544,] 0.8754310
#>  [545,] 0.7113970
#>  [546,] 1.0043579
#>  [547,] 0.8690696
#>  [548,] 0.9600325
#>  [549,] 1.2121802
#>  [550,] 0.9635409
#>  [551,] 0.7831222
#>  [552,] 1.2298517
#>  [553,] 0.9191643
#>  [554,] 1.0239431
#>  [555,] 1.0589733
#>  [556,] 1.0202987
#>  [557,] 0.8850015
#>  [558,] 0.9947683
#>  [559,] 0.8900164
#>  [560,] 0.7449006
#>  [561,] 0.9616550
#>  [562,] 1.1509015
#>  [563,] 0.8632563
#>  [564,] 1.0350111
#>  [565,] 0.9845858
#>  [566,] 0.9989405
#>  [567,] 1.1145548
#>  [568,] 1.1585761
#>  [569,] 0.9227044
#>  [570,] 0.8321092
#>  [571,] 0.8180654
#>  [572,] 1.0342469
#>  [573,] 1.0673413
#>  [574,] 1.3217427
#>  [575,] 1.0041777
#>  [576,] 0.9721153
#>  [577,] 1.0856138
#>  [578,] 0.8969349
#>  [579,] 1.1296122
#>  [580,] 1.0904709
#>  [581,] 1.1409165
#>  [582,] 0.9541442
#>  [583,] 1.1384222
#>  [584,] 0.9296453
#>  [585,] 0.9815878
#>  [586,] 1.1310674
#>  [587,] 0.8242785
#>  [588,] 0.9402007
#>  [589,] 0.9302146
#>  [590,] 0.9695138
#>  [591,] 0.8728202
#>  [592,] 1.0737293
#>  [593,] 0.9206213
#>  [594,] 0.8118484
#>  [595,] 0.8655899
#>  [596,] 1.2235222
#>  [597,] 1.1345478
#>  [598,] 0.8261419
#>  [599,] 1.1100462
#>  [600,] 1.0727239
#>  [601,] 0.9990929
#>  [602,] 1.0830517
#>  [603,] 1.0033982
#>  [604,] 0.8744400
#>  [605,] 0.8169172
#>  [606,] 1.3878243
#>  [607,] 0.8501468
#>  [608,] 0.8686654
#>  [609,] 1.1479124
#>  [610,] 0.9134748
#>  [611,] 1.0308013
#>  [612,] 0.9116965
#>  [613,] 0.7568714
#>  [614,] 0.9983845
#>  [615,] 1.1959727
#>  [616,] 1.0241877
#>  [617,] 1.2440421
#>  [618,] 1.2077349
#>  [619,] 0.9478244
#>  [620,] 1.3913398
#>  [621,] 0.9962673
#>  [622,] 0.9182902
#>  [623,] 0.9939202
#>  [624,] 0.8240042
#>  [625,] 1.2281542
#>  [626,] 1.0005217
#>  [627,] 0.9330081
#>  [628,] 0.9093467
#>  [629,] 1.1340139
#>  [630,] 0.8691090
#>  [631,] 1.0573735
#>  [632,] 1.1543958
#>  [633,] 1.1296158
#>  [634,] 1.2430400
#>  [635,] 1.0430623
#>  [636,] 1.1357716
#>  [637,] 0.8323500
#>  [638,] 0.9187748
#>  [639,] 1.1422276
#>  [640,] 0.8879610
#>  [641,] 1.1403040
#>  [642,] 1.2917563
#>  [643,] 0.8636800
#>  [644,] 0.9169678
#>  [645,] 0.9990779
#>  [646,] 1.1062363
#>  [647,] 1.4012648
#>  [648,] 0.8892082
#>  [649,] 1.1374096
#>  [650,] 0.7962698
#>  [651,] 0.9336621
#>  [652,] 0.8303555
#>  [653,] 1.0748241
#>  [654,] 0.9838151
#>  [655,] 1.0644647
#>  [656,] 0.9850535
#>  [657,] 1.0498456
#>  [658,] 0.8685121
#>  [659,] 0.9061638
#>  [660,] 1.1386658
#>  [661,] 0.9289590
#>  [662,] 0.9532943
#>  [663,] 1.1989537
#>  [664,] 0.9882093
#>  [665,] 0.9000640
#>  [666,] 1.1016159
#>  [667,] 1.0818959
#>  [668,] 0.8483761
#>  [669,] 0.9581510
#>  [670,] 1.0847278
#>  [671,] 1.0166109
#>  [672,] 1.1106329
#>  [673,] 1.1386653
#>  [674,] 1.1577557
#>  [675,] 0.8717209
#>  [676,] 0.8249461
#>  [677,] 1.0077017
#>  [678,] 0.8245639
#>  [679,] 1.2722653
#>  [680,] 0.7284762
#>  [681,] 1.2845075
#>  [682,] 0.9876730
#>  [683,] 0.8625741
#>  [684,] 0.9928540
#>  [685,] 0.8785424
#>  [686,] 0.9415048
#>  [687,] 1.2711642
#>  [688,] 0.9417822
#>  [689,] 0.9816644
#>  [690,] 1.1198899
#>  [691,] 0.9587339
#>  [692,] 1.2750232
#>  [693,] 1.2481629
#>  [694,] 0.8438466
#>  [695,] 1.1781945
#>  [696,] 0.9122764
#>  [697,] 1.3679662
#>  [698,] 1.0691674
#>  [699,] 0.9178546
#>  [700,] 0.9026894
#>  [701,] 0.9691144
#>  [702,] 0.8835672
#>  [703,] 0.8596459
#>  [704,] 0.9925180
#>  [705,] 0.9841153
#>  [706,] 0.9636690
#>  [707,] 0.8945521
#>  [708,] 0.8989720
#>  [709,] 0.8314678
#>  [710,] 0.9261598
#>  [711,] 1.0210824
#>  [712,] 0.7786989
#>  [713,] 0.9217183
#>  [714,] 0.9187307
#>  [715,] 0.8336838
#>  [716,] 1.2164967
#>  [717,] 1.1093400
#>  [718,] 0.9974363
#>  [719,] 1.3216062
#>  [720,] 0.7273483
#>  [721,] 0.8346694
#>  [722,] 0.8056654
#>  [723,] 0.8902586
#>  [724,] 0.9289086
#>  [725,] 1.0074691
#>  [726,] 0.9295779
#>  [727,] 0.8903845
#>  [728,] 0.8187074
#>  [729,] 0.9434709
#>  [730,] 0.8503582
#>  [731,] 1.2790916
#>  [732,] 1.1064403
#>  [733,] 0.7908667
#>  [734,] 0.9639611
#>  [735,] 1.0941133
#>  [736,] 1.0477527
#>  [737,] 0.9190737
#>  [738,] 0.9679257
#>  [739,] 0.8167906
#>  [740,] 1.0797071
#>  [741,] 1.2332580
#>  [742,] 1.0063514
#>  [743,] 0.9656280
#>  [744,] 0.9282617
#>  [745,] 0.9198348
#>  [746,] 0.9611282
#>  [747,] 0.9419518
#>  [748,] 0.7713708
#>  [749,] 1.1588342
#>  [750,] 0.9896375
#>  [751,] 0.9452921
#>  [752,] 0.9798264
#>  [753,] 1.1083243
#>  [754,] 0.7550860
#>  [755,] 0.8191351
#>  [756,] 1.0994988
#>  [757,] 1.0325167
#>  [758,] 0.6756269
#>  [759,] 1.0617643
#>  [760,] 0.9389387
#>  [761,] 1.0747293
#>  [762,] 1.0878533
#>  [763,] 0.9128124
#>  [764,] 0.8376191
#>  [765,] 1.1948393
#>  [766,] 0.8654037
#>  [767,] 0.8651701
#>  [768,] 0.7688215
#>  [769,] 1.1583657
#>  [770,] 1.0509560
#>  [771,] 1.1381946
#>  [772,] 1.0162302
#>  [773,] 1.0275672
#>  [774,] 0.7907422
#>  [775,] 0.9472122
#>  [776,] 0.8853425
#>  [777,] 1.2253041
#>  [778,] 0.7820729
#>  [779,] 1.0118702
#>  [780,] 1.0171566
#>  [781,] 0.8752778
#>  [782,] 0.8081004
#>  [783,] 0.9670309
#>  [784,] 0.9984016
#>  [785,] 0.9541673
#>  [786,] 0.9598537
#>  [787,] 0.8857177
#>  [788,] 0.8256411
#>  [789,] 0.9246664
#>  [790,] 0.9778895
#>  [791,] 1.1483442
#>  [792,] 0.9977355
#>  [793,] 1.1510481
#>  [794,] 1.0234894
#>  [795,] 0.9794956
#>  [796,] 1.1560212
#>  [797,] 1.1828419
#>  [798,] 0.9641353
#>  [799,] 0.9416289
#>  [800,] 1.0700007
#>  [801,] 0.8811105
#>  [802,] 0.9381136
#>  [803,] 0.9812740
#>  [804,] 0.9928310
#>  [805,] 0.9684861
#>  [806,] 0.9871281
#>  [807,] 1.0615431
#>  [808,] 1.0332128
#>  [809,] 1.0762488
#>  [810,] 1.0887231
#>  [811,] 0.7563552
#>  [812,] 1.0637631
#>  [813,] 0.7750646
#>  [814,] 1.1478851
#>  [815,] 1.1596067
#>  [816,] 0.9244504
#>  [817,] 1.2776775
#>  [818,] 0.8294132
#>  [819,] 0.9235941
#>  [820,] 0.9354148
#>  [821,] 0.8961944
#>  [822,] 0.7567138
#>  [823,] 0.9072347
#>  [824,] 0.7715338
#>  [825,] 0.9320888
#>  [826,] 1.1125741
#>  [827,] 1.1325455
#>  [828,] 0.8174643
#>  [829,] 0.8490710
#>  [830,] 0.9970368
#>  [831,] 1.0990970
#>  [832,] 1.0516287
#>  [833,] 1.0906604
#>  [834,] 0.7982365
#>  [835,] 0.8797601
#>  [836,] 0.9181764
#>  [837,] 0.9675407
#>  [838,] 1.0792681
#>  [839,] 1.0170387
#>  [840,] 1.1185055
#>  [841,] 1.2291185
#>  [842,] 0.9087035
#>  [843,] 0.8340705
#>  [844,] 0.8180600
#>  [845,] 0.7364960
#>  [846,] 0.9197248
#>  [847,] 1.2716538
#>  [848,] 0.9503990
#>  [849,] 0.7937433
#>  [850,] 0.8604354
#>  [851,] 1.0601606
#>  [852,] 0.9831686
#>  [853,] 0.8994062
#>  [854,] 0.9632143
#>  [855,] 0.8946500
#>  [856,] 1.0596044
#>  [857,] 0.9153416
#>  [858,] 0.9450553
#>  [859,] 0.9976198
#>  [860,] 1.0501875
#>  [861,] 1.4309402
#>  [862,] 0.8752931
#>  [863,] 1.0002249
#>  [864,] 1.2310320
#>  [865,] 1.0049840
#>  [866,] 1.1918087
#>  [867,] 0.8302866
#>  [868,] 1.0116910
#>  [869,] 0.9537322
#>  [870,] 1.0687661
#>  [871,] 0.9954617
#>  [872,] 1.1419255
#>  [873,] 0.8528384
#>  [874,] 0.9719062
#>  [875,] 1.0669039
#>  [876,] 1.1255419
#>  [877,] 1.0045404
#>  [878,] 1.0613506
#>  [879,] 0.9806485
#>  [880,] 1.0531218
#>  [881,] 0.9042595
#>  [882,] 0.8176620
#>  [883,] 0.8553269
#>  [884,] 0.9055752
#>  [885,] 0.9879487
#>  [886,] 1.2481913
#>  [887,] 0.8500857
#>  [888,] 0.8309833
#>  [889,] 1.1104185
#>  [890,] 1.1177713
#>  [891,] 0.9550683
#>  [892,] 0.8303140
#>  [893,] 1.2566292
#>  [894,] 1.1506101
#>  [895,] 1.0310631
#>  [896,] 1.0846391
#>  [897,] 1.1391169
#>  [898,] 0.7794826
#>  [899,] 1.1009652
#>  [900,] 0.7948389
#>  [901,] 0.8564466
#>  [902,] 0.9849936
#>  [903,] 1.1083243
#>  [904,] 1.0599351
#>  [905,] 0.9222007
#>  [906,] 1.0356712
#>  [907,] 1.0182542
#>  [908,] 1.0763449
#>  [909,] 1.2775636
#>  [910,] 0.9951745
#>  [911,] 1.0146841
#>  [912,] 0.9745596
#>  [913,] 0.9154685
#>  [914,] 0.8457953
#>  [915,] 0.9275279
#>  [916,] 0.9452668
#>  [917,] 0.8926834
#>  [918,] 1.0717918
#>  [919,] 0.9908221
#>  [920,] 0.8306369
#>  [921,] 1.0356706
#>  [922,] 0.7605846
#>  [923,] 1.0047124
#>  [924,] 1.0659615
#>  [925,] 1.2108448
#>  [926,] 0.9468451
#>  [927,] 0.8544984
#>  [928,] 1.2177416
#>  [929,] 0.7344766
#>  [930,] 0.9358432
#>  [931,] 0.8620742
#>  [932,] 1.2466800
#>  [933,] 1.0389866
#>  [934,] 1.0341162
#>  [935,] 1.1416042
#>  [936,] 0.9281608
#>  [937,] 0.8220133
#>  [938,] 0.8978391
#>  [939,] 1.0143270
#>  [940,] 1.0139655
#>  [941,] 0.9808338
#>  [942,] 0.7790992
#>  [943,] 0.9266748
#>  [944,] 0.8342006
#>  [945,] 1.0577773
#>  [946,] 0.9360060
#>  [947,] 1.0395118
#>  [948,] 1.0344709
#>  [949,] 0.9749826
#>  [950,] 1.2225973
#>  [951,] 0.8165587
#>  [952,] 0.8929384
#>  [953,] 0.8991947
#>  [954,] 0.8723987
#>  [955,] 1.1389572
#>  [956,] 0.9281150
#>  [957,] 1.0372396
#>  [958,] 1.4120541
#>  [959,] 0.9053824
#>  [960,] 1.0915218
#>  [961,] 0.8123744
#>  [962,] 1.0425235
#>  [963,] 0.8981404
#>  [964,] 1.0825070
#>  [965,] 0.9598202
#>  [966,] 0.8407516
#>  [967,] 0.9019217
#>  [968,] 0.8105849
#>  [969,] 1.0809844
#>  [970,] 0.8782786
#>  [971,] 0.9593084
#>  [972,] 0.9662043
#>  [973,] 0.9284369
#>  [974,] 0.9413302
#>  [975,] 1.1060359
#>  [976,] 1.1576839
#>  [977,] 1.0316770
#>  [978,] 0.9422652
#>  [979,] 0.9519493
#>  [980,] 0.7684105
#>  [981,] 1.0700721
#>  [982,] 1.2314079
#>  [983,] 0.9952335
#>  [984,] 1.1939661
#>  [985,] 1.0779970
#>  [986,] 1.3047445
#>  [987,] 0.9891054
#>  [988,] 0.8980994
#>  [989,] 1.0667042
#>  [990,] 0.9492103
#>  [991,] 1.0042490
#>  [992,] 0.8543314
#>  [993,] 0.8200392
#>  [994,] 0.9156050
#>  [995,] 0.8300312
#>  [996,] 0.8606008
#>  [997,] 1.0783280
#>  [998,] 1.1951833
#>  [999,] 0.9876522
#> [1000,] 0.9573271
# }