Computes predicted dominance phase durations using posterior predictive distribution.

Computes predicted dominance phase durations using fitted model. Returns predicted values only for the dominance phases that were marked for use. I.e., excluding first and last dominance phases, mixed phases, etc. See preprocess_data().

Usage

# S3 method for cumhist
predict(object, summary = TRUE, probs = NULL, full_length = TRUE, ...)

Arguments

object

An object of class cumhist

summary

Whether summary statistics should be returned instead of raw sample values. Defaults to TRUE

probs

The percentiles used to compute summary, defaults to NULL (no CI).

full_length

Only for summary = TRUE, whether the summary table should include rows with no predictions. I.e., rows with mixed phases, first/last dominance phase in the run, etc. See preprocess_data(). Defaults to TRUE.

...

Unused

Value

If summary=FALSE, a numeric matrix iterationsN x clearN. If summary=TRUE but probs=NULL a vector of mean predicted durations. If summary=TRUE and probs is not NULL, a data.frame with a column "Predicted" (mean) and a column for each specified quantile.

See also

Examples

# \donttest{
br_fit <- fit_cumhist(br_singleblock, state = "State", duration = "Duration")
#> 
#> SAMPLING FOR MODEL 'historylm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 5.106 seconds (Warm-up)
#> Chain 1:                5.898 seconds (Sampling)
#> Chain 1:                11.004 seconds (Total)
#> Chain 1: 
predict(br_fit)
#>  [1]       NA       NA 3.108167 3.031731 3.103224 4.978781 3.556191 4.391068
#>  [9] 3.323266 3.693118 4.186118 5.258083 4.152798 4.602051 3.813456 3.976780
#> [17] 4.107488 3.409408 3.863495 3.567847 4.048410 3.810553 4.390700 4.468915
#> [25] 3.884608 3.801234 5.092894 4.192921 5.188636 4.865705 3.567865 4.383723
#> [33] 3.821524 4.366296 4.815683 3.693076 4.360781 5.417178 4.445433 4.040598
#> [41] 2.959999 2.850569 4.373591 5.083916 3.798222 5.219255 2.597284       NA
#> [49] 2.532357       NA 2.608537 3.202247 4.190415 3.707502 5.145955 5.646412
#> [57] 3.602320       NA 2.352294 3.240449 4.477335 3.022112 3.683759 4.597077
#> [65] 4.707207 3.279707 3.722988 4.794496 4.727354 4.900426 5.622265 5.563394
#> [73] 3.975545 4.423519 4.667857       NA

# full posterior prediction samples
predictions_samples <- predict(br_fit, summary=FALSE)
# }